requirements.txt
streamlit
zhipuai
pypdf
python-dotenv
pandas
